Complete rewrite of project
Made project more scalable, added some dep management features, but removed most as it's unmaintainable. Removed ffmpeg requirement.
