+++ .gitignore
@@ -1,2 +1,3 @@
 test.py
 .vscode
+.idea
