+++ requirements.txt
@@ -1,14 +1,5 @@
-# awsfunc Deps, awsfunc is on self hosted git and not in the pip repos. awsfunc frozen 
-VapourSynth>=49
-vsutil@https://github.com/Irrational-Encoding-Wizardry/vsutil/archive/08d782115a7464e238a38ea836a44571bcfeecc3.zip
-rekt@https://gitlab.com/Ututu/rekt/-/archive/3da2b2f2b2d670e635cef6dcc61f19c8fe10f1fa/rekt-3da2b2f2b2d670e635cef6dcc61f19c8fe10f1fa.zip
-
-# Package is abandoned
-screenutils
-
-vsutil
-numpy
-opencv-python
-
-#Package is abandoned
-termcolor
+awsmfunc>=1.3.3
+numpy>=1.24.0
+opencv-contrib-python>=4.6.0.68
+VapourSynth>=60
+vsutil>=0.8.0
